requirements.txt
Flask==2.2.2
pymongo==4.3.3
python-dotenv==0.21.1
graphene==2.1.9
bcrypt==4.0.1
PyJWT==2.8.0
flask-graphql==2.0.1